05_canSniffer_Flutter/api_server/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-cors==3.0.10
pyserial==3.5
python-engineio==4.2.1
python-socketio==5.4.0
werkzeug==2.0.1
gevent==21.8.0
gevent-websocket==0.10.1